+++ CHANGELOG.md
@@ -11,12 +11,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 * [#249](https://github.com/shlinkio/shlink-web-component/issues/249) Replace `react-datepicker` with native `input[type="date"]` and `input[type="datetime-local"]` elements.
 * [#257](https://github.com/shlinkio/shlink-web-component/issues/257) Remove dependency on `react-copy-to-clipboard`.
+* [#278](https://github.com/shlinkio/shlink-web-component/issues/278) Update to `@shlinkio/shlink-js-sdk` v1.0.0.
 
 ### Deprecated
 * *Nothing*
 
 ### Removed
-* *Nothing*
+* [#276](https://github.com/shlinkio/shlink-web-component/issues/276) Drop support for Shlink older than 3.3.0.
 
 ### Fixed
 * *Nothing*